pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "network-sniffer"
version = "1.0.0"
description = "Enterprise Network Diagnostic Tool - Real-time packet capture and analysis"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Network Team"}
]
keywords = ["network", "sniffer", "packet-capture", "diagnostics", "monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "scapy>=2.5.0",
    "textual>=0.40.0",
    "rich>=13.0.0",
    "plotext>=5.0.0",
    "numpy>=1.24.0",
    "psutil>=5.9.0",
    "netifaces>=0.11.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
yaml = [
    "pyyaml>=6.0.0",
]
all = [
    "pyyaml>=6.0.0",
]

[project.scripts]
network-sniffer = "network_sniffer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["network_sniffer*"]
